+++ src/features/auth/components/logout-button.tsx
@@ -1,0 +1,15 @@
+export const LogoutButton = () => {
+  const handleLogout = async () => {
+    await fetch(`/api/auth/logout`, {
+      method: "POST",
+    });
+
+    window.location.href = "/";
+  };
+
+  return (
+    <button class="cursor-pointer" onClick={handleLogout}>
+      Logout
+    </button>
+  );
+};