+++ src/middleware.ts
@@ -9,8 +9,6 @@ export const onRequest = defineMiddleware(async (context, next) => {
   const sessionToken = context.cookies.get(AUTH_COOKIES.SESSION_TOKEN)?.value;
   const sessionInfo = await getSessionInfo(sessionToken);
 
-  console.log("sessionInfo", sessionInfo);
-
   if (sessionInfo && sessionInfo.user) {
     context.locals.userId = sessionInfo.user.id;
     // Extend session if it's about to expire